### 2025-05-17 #8217 @editorialbot
[PRE REVIEW]: An R Package for Acquiring and Processing Notifiable Infectious Diseases Dataset from the Japan Institute for Health Security
**Submitting author:** @TomonoriHoshi (<a href="http://orcid.org/0000-0003-4569-4600">Tomonori Hoshi</a>)
**Repository:** https://github.com/TomonoriHoshi/jpinfect
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.0
