### 2025-05-20 0b42aaa
DOI URL errors fixed
