+++ joss/paper.bib
@@ -26,7 +26,7 @@
   number = {7},
   pages = {868-882},
   keywords = {biostatistics and bioinformatics, coronavirus, data processing, epidemiology, infection, test statistics, virus classification},
-  doi = {https://doi.org/10.1002/jmv.25885},
+  doi = {10.1002/jmv.25885},
   url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/jmv.25885},
   eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/jmv.25885},
   year = {2020}
@@ -68,7 +68,7 @@
   pages = {325-337},
   year = {2016},
   issn = {0740-624X},
-  doi = {https://doi.org/10.1016/j.giq.2016.02.001},
+  doi = {10.1016/j.giq.2016.02.001},
   url = {https://www.sciencedirect.com/science/article/pii/S0740624X16300132},
   author = {Antonio Vetrò and Lorenzo Canova and Marco Torchiano and Camilo Orozco Minotas and Raimondo Iemma and Federico Morando},
   keywords = {Open Government Data, Open data quality, Government information quality, Data quality measurement, Empirical assessment},
