+++ DESCRIPTION
@@ -22,7 +22,7 @@ Authors@R: c(
 Maintainer: Tomonori Hoshi <tomonori.hoshi.japan@gmail.com>
 Description: Functions to download and post process the infectious disease case data 
               from Japan Institute for Health Security.
-License: GPL (>= 3)
+License: GPL (>= 3) + file LICENSE
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 7.3.2
