### 2025-05-24 f55593e
Description improved to follow CRAN
