+++ DESCRIPTION
@@ -21,8 +21,10 @@ Authors@R: c(
   )
 Maintainer: Tomonori Hoshi <tomonori.hoshi.japan@gmail.com>
 Description: Functions to download and post process the infectious disease case data 
-              from Japan Institute for Health Security.
-License: GPL (>= 3) + file LICENSE
+              from Japan Institute for Health Security (JIHS). Also the package included
+              ready-to-analyse dataset from JIHS. See JISH website for further details
+              <https://id-info.jihs.go.jp/usage-contract.html>.
+License: GPL (>= 3)
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 7.3.2
