+++ vignettes/jpinfect.Rmd
@@ -61,12 +61,12 @@ The `jpinfect` package depends on the following R packages:
 
 ## Installation
 
-The `jpinfect` package can be installed from GitHub using the [devtools](https://github.com/hadley/devtools) package. To install the package, run the following command in your R console:
+The `jpinfect` package can be installed from GitHub using the [remotes](https://github.com/r-lib/remotes) package. To install the package, run the following command in your R console:
 
 ```{r, message = FALSE, warning = FALSE, eval = FALSE}
 # install.packages("jpinfect")
-if(!require("devtools")) install.packages("devtools")
-devtools::install_github("TomonoriHoshi/jpinfect")
+if(!require("remotes")) install.packages("remotes")
+remotes::install_github("TomonoriHoshi/jpinfect")
 ```
 
 Load the package after installation:
