### 2025-05-29 996423d
Updates examples and dontruns to follow CRAN's guideline
