+++ R/jpinfect_url_bullet.R
@@ -25,7 +25,7 @@
 #' @importFrom httr HEAD
 #'
 #' @examples
-#' \dontrun{
+#' \donttest{
 #' # Display URLs for English data for weeks 1 to 5 in 2025
 #' jpinfect_url_bullet(year = 2025, week = 1:5, language = "en")
 #'
