+++ R/jpinfect_pivot.R
@@ -21,7 +21,7 @@
 #' - The function ensures that the columns `prefecture`, `year`, `week` and `date` are always present.
 #'
 #' @examples
-#' \dontrun{
+#' \donttest{
 #' # Pivot the dataset
 #' result <- jpinfect_pivot(bind_dataset)
 #' print(result)
