+++ man/jpinfect_pivot.Rd
@@ -30,7 +30,7 @@ it is pivoted to a long format.
 \examples{
 \donttest{
 # Pivot the dataset
-result <- jpinfect_pivot(bind_dataset)
+result <- jpinfect_pivot(place_prefecture)
 print(result)
 }
 
