### 2025-06-05 3ee161e
Readme updates for Cran installation
