### 2025-07-09 6d1b994
grand-total dl badges update
