### 2025-07-10 #8217 @rkurchin
(I know this is a bit outside your scope, but it's been waiting for awhile and the deep technical details are really the job of the reviewers 😄)
