### 2025-07-10 #8217 @fboehm
@joseph-palmer @mponce0 @federicomarini @ArdiaD @LogIN- @JDRomano2 - Would you be willing to review this submission for JOSS? As you can see, it's for an R package. Thank you for considering this invitation.
