### 2025-07-10 #8217 @fboehm
@eguidotti - would you be willing to review this manuscript for JOSS? Our reviews are checklist-driven. Thank you for considering this invitation.
