### 2025-07-14 7976611
Fixed descriptions on data duration and  folder settings for data downloads
