### 2025-08-03 58e56d4
Detailed description for manual fixed and version updated to 0.1.3.
