+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: jpinfect
 Type: Package
 Title: Acquiring and Processing Data from Japan Institute for Health Security
-Version: 0.1.2
+Version: 0.1.3
 Authors@R: c(
     person(
       "Tomonori", "Hoshi",
