### 2025-08-05 #8577 @mponce0
# General Checks
## repository
    - some improvement in terms of organization and overall presentation would be encourage.
        E.g. in the R directory there is a zzz.R file which is named not in a
        "conventional" fashion or following usual standards, more likely a
        left-over.

    - the README document includes a section in how to report bugs,
    I would also suggest the addition of a "Contribution" document/section,
    which could be useful for interested parties, detailing the procedural aspects of
    potential contributions to the package

# Software paper
    - Quality of Writing:
        (minor) writing can be improved and may need revisiting to improve continuity and flow of reading

    - State of the field
        there is no comparison to other packages or the actual process of obtaining the data directly
        from the Japan Institute for Health Security -- this shows up below but a more enlightening
        comparison would be useful

# Functionality
    - installation
        - tested on Linux Ubuntu
            - worked for CRAN & GitHub
            - did not test multi-OSes but CRAN checks for that

    - functionality
        * Data Acquisition & Data Import appear to be FAILING!
        - some of the examples in "Data Acquisition" session appear to be failing
        or at least displaying some warning/errors at the beginning although the download appear to carry on
        It may be worth checking this or providing more detailed information to the user about this.
        - Hence, the examples under "Data Import" are failing.

    - performance
        - in the paper there is the claim that the package provides an "*efficient* and streamlined pipeline"
        but not real justification or argument about it is stated; there is room for justifying or explaining this further

# Documentation
    - Automated Tests
        working OK, but only way to trigger them were by using devtools::test() in the package folder
        shouldn't be an issue as the package is part of CRAN

    - Examples usage
        the examples as shown in the GitHub repository are quite basic and some of them FAILING;
        a full example of real case application would be ideal

    - Community guidelines
        as mentioned before, I feel that the authors should include further details about
        possible contributors as pull requests, issues in the repository, etc.

-----
