### 2025-08-13 #8577 @TomonoriHoshi
@fboehm @JDRomano2 @mponce0

Thank you very much for handling our submission and review.   Our apologies for the command errors in the instructions. Now, we identified the issues.  Once all reviews are complete, the corrected code will be push to "[joss-review-response](https://github.com/TomonoriHoshi/jpinfect/tree/joss-review-response)" branch.
