### 2025-08-18 a299bff
docs: create CONTRIBUTING.md and restructure README.md
Moved "Report a bug" section to CONTRIBUTING.md in response to @mponce0's comments. Added "Community guidelines" to README.md to streamline content.

#####
- the README document includes a section in how to report bugs,
I would also suggest the addition of a "Contribution" document/section,
which could be useful for interested parties, detailing the procedural aspects of
potential contributions to the package
