### 2025-08-18 a7dd1ef
AI disclosure sentence added as suggested by @mponce0
#######
AI disclosure
- no details about the use of AI technology is mentioned
