### 2025-08-18 3ffb959
Add real case example for epidemiologists (per @mponce0's suggestion)
########
- Examples usage
    the examples as shown in the GitHub repository are quite basic and some of them FAILING;
    a full example of real case application would be ideal
