### 2025-09-11 fb7074c
Improve software paper following @mponce0 suggestions. Especially, the "Statement of the field" section has been enhanced.
- Quality of Writing:
(minor) writing can be improved and may need revisiting to improve continuity and flow of reading

- State of the field
there is no comparison to other packages or the actual process of obtaining the data directly from the Japan Institute for Health Security -- this shows up below but a more enlightening comparison would be useful

- Functionality - performance
- in the paper there is the claim that the package provides an "*efficient* and streamlined
[…]
