+++ joss/paper.md
@@ -22,7 +22,7 @@ authors:
   affiliation: 1, 2, 4, 5
 bibliography: paper.bib
 affiliations:
-- name: Institute of Tropical Medicine, Nagasaki University
+- name: Institute of Tropical Medicine, Nagasaki University, Japan
   index: 1
 - name: School of Tropical Medicine and Global Health, Nagasaki University, Japan
   index: 2
@@ -36,23 +36,23 @@ affiliations:
 
 # Summary
 
-`jpinfect` is an R [@Rdev:2025] package provides a set of functions to acquire and post-process notifiable infectious disease datasets from the Japan Institute for Health Security [@JIHS]. The package facilitates to generate combined datasets of weekly case reports since week 14th 1999 by prefecture, where available, sex and suspected location of infection information. In addition to its core functionalities, the package also includes built-in datasets that have already been pre-processed using the provided functions. These datasets are ready for immediate analysis, making it easier for researchers, public health practitioners and educators to utilise officially released public data. The package is designed to streamline epidemiological research, enhance public health response and support educational efforts. Ultimately, `jpinfect` aims to assist researchers and practitioners in responding to notifiable infectious diseases in Japan efficiently. All package code is archived on [GitHub](https://github.com/TomonoriHoshi/jpinfect), allowing users to extend its functionality and adapt it to their specific needs.
+`jpinfect` is an R [@Rdev:2025] package that provides a set of functions to acquire and process notifiable infectious disease datasets from the Japan Institute for Health Security [@JIHS]. The package helps generate combined datasets of weekly case reports since week 14 of 1999 by prefecture and, where available, sex and suspected location of infection information. In addition to its core functionalities, the package also includes pre-processed built-in datasets. These datasets are ready for immediate analysis, making it easier to utilise officially released public data. The package is designed to streamline epidemiological research, enhance public health response and support educational efforts. Ultimately, `jpinfect` aims to assist researchers and practitioners in responding efficiently to notifiable infectious diseases in Japan. All code is archived on [GitHub](https://github.com/TomonoriHoshi/jpinfect) for extension and adaptation.
 
 # Statement of need
 
-The global pandemic of COVID-19 highlighted the importance of reliable epidemiological dataset for outbreak monitoring and effective disease control [@Dehkordi:2020]. Thanks to computer advancements, mathematical modelling methods based on the large dataset could be implemented in many countries for the COVID-19 response [@Vynnycky:2011; @Ferguson:2020]. This experience emphasised that timely and accurate information is one of the most important keys for successful disease control [@Gulland:2020]. Indeed, public health officers and epidemiologists worked for governmental organisations paid their greatest attention about whether any epidemiological suggestions were based on the reliable data sources [@WHO:2020].
+The COVID-19 pandemic highlighted the importance of reliable epidemiological dataset for outbreak monitoring and effective disease control [@Dehkordi:2020]. Thanks to computer advancements, mathematical modelling methods based on the large dataset could be implemented in many countries for the COVID-19 response [@Vynnycky:2011; @Ferguson:2020]. This experience emphasised that timely and accurate information is crucial for successful disease control [@Gulland:2020]. Consequently, public health officers and epidemiologists at government organisations paid particular attention to ensuring that all epidemiological recommendations were grounded in reliable data sources [@WHO:2020].
 
-In real-world settings, officially released datasets from government organisations, while comprehensive, are the most reliable source. However, these are often difficult to process for immediate use due to their complex structure [@Vetrò:2016]. This barrier hindered rapid response for containing the outbreak [@Stoto:2022]. The ability to prepare, clean, and standardise these datasets during non-emergency periods is essential for improving responsiveness in the event of future outbreaks.
[…]
