+++ joss/paper.bib
@@ -121,3 +121,56 @@
 	url = {https://www.i-jmr.org/2023/1/e44310},
 	langid = {en}
 }
+
+
+@article{Hanscombe:2019,
+    title = {ukbtools: An R package to manage and query UK Biobank data},
+    author = {Hanscombe, Ken B. AND Coleman, Jonathan R. I. AND Traylor, Matthew AND Lewis, Cathryn M.},
+    year = {2019},
+    month = {05},
+    journal = {PLOS ONE},
+    publisher = {Public Library of Science},
+    doi = {10.1371/journal.pone.0214311},
+    volume = {14},
+    url = {https://doi.org/10.1371/journal.pone.0214311},
+    pages = {1-6},
+    number = {5},
+}
+
+
+@Manual{epidm,
+  title        = {epidm: UK Epidemiological Data Management},
+  author       = {Alex Bhattacharya},
+  year         = {2022},
+  note         = {R package version 1.0.4},
+  doi          = {10.32614/CRAN.package.epidm},
+  url          = {https://CRAN.R-project.org/package=epidm}
+}
+
+@Manual{epidatr,
+  title        = {epidatr: Client for Delphi's 'Epidata' API},
+  author       = {Logan Brooks and Dmitry Shemetov and Samuel Gratzl and David Weber and Nat DeFries and Alex Reinhart and Daniel J. McDonald and Kean Ming Tan and Will Townes and George Haff and Kathryn Mazaitis},
+  year         = {2025},
+  note         = {R package version 1.2.1},
+  doi          = {10.32614/CRAN.package.epidatr},
+  url          = {https://CRAN.R-project.org/package=epidatr}
+}
+
+
+@Misc{cdcfluview,
+  title        = {cdcfluview: Retrieve Flu Season Data from the United States Centers for Disease Control and Prevention ('CDC') 'FluView' Portal},
+  author       = {Bob Rudis and Craig McGowan and JJ Chen and Sebastian Meyer and James Turtle and Andrew Bates and Ian McGovern},
+  year         = {2022},
+  note         = {R package version 1.0.0.9000, GitHub repository},
+  url          = {https://github.com/hrbrmstr/cdcfluview}
+}
+
+
+@Manual{covidcast,
+  title        = {covidcast: Client for Delphi's 'COVIDcast Epidata' API},
+  author       = {Taylor Arnold and Jacob Bien and Logan Brooks and Sarah Colquhoun and David Farrow and Jed Grabman and Pedrito Maynard-Zhang and Kathryn Mazaitis and Alex Reinhart and Ryan Tibshirani},
+  year         = {2025},
+  note         = {R package version 0.5.3},
+  doi          = {10.32614/CRAN.package.covidcast},
+  url          = {https://CRAN.R-project.org/package=covidcast}
+}
