### 2025-09-17 4758018
Add GitHub Actions workflow for R-CMD-check (check-standard), a bit adjustment for the develop branch
