### 2025-09-18 TomonoriHoshi/jpinfect#6 @mponce0
"Data Acquisition" & "Data Import" examples
# Data Acquisition examples
when running the example for "Data Acquisition" although the example
continue executing it also produces some "confusing" warning messages, e.g.

```
# Pivot the dataset: wider to longer
bullet_long <- jpinfect_pivot(bullet)

# Pivot the dataset: longer to wider
bullet_wide <-  jpinfect_pivot(bullet_long)
>
> ## Check the source URL for confirmed dataset
# Download case reports by sex and prefecture
jpinfect_url_confirmed(year = 2021, type = "sex")

# Download case reports by the place of incetion and prefecture
jpinfect_url_confirmed(year = 2021, type = "place")
[1] "https://id-info.jihs.go.jp/surveillance/idwr/annual/2021/syulist/Syu_01_1.xlsx"
[1] "https://id-info.jihs.go.jp/surveillance/idwr/annual/2021/syulist/Syu_02_1.xlsx"
>
> # Download data for 2020 and 2021
jpinfect_get_confirmed(years = c(2020, 2021), type = "sex", dest_dir = "raw_data")

# Download data for all available years
jpinfect_get_confirmed(type = "place", dest_dir = "raw_data")

# Download English data for weeks 1 to 5 in 2025
jpinfect_get_bullet(year = 2025, week = 1:5, dest_dir = "raw_data")

# Download Japanese data for all weeks in 2025
jpinfect_get_bullet(year = 2025, language = "jp", dest_dir = "raw_data")
Directory 'raw_data' does not exist. Please create it manually.
NULL
Directory 'raw_data' does not exist. Please create it manually.
NULL
All requested weeks have been processed successfully.
Downloading data for week 1 from URL: https://id-info.jihs.go.jp/surveillance/idwr/en/rapid/2025/01/zensu01.csv
Download completed: raw_data/2025_01_zensu_en.csv
Downloading data for week 2 from URL: https://id-info.jihs.go.jp/surveillance/idwr/en/rapid/2025/02/zensu02.csv
Download completed: raw_data/2025_02_zensu_en.csv
Downloading data for week 3 from URL: https://id-info.jihs.go.jp/surveillance/idwr/en/rapid/2025/03/zensu03.csv
Download completed: raw_data/2025_03_zensu_en.csv
Downloading data for week 4 from URL: https://id-info.jihs.go.jp/surveillance/idwr/en/rapid/2025/04/zensu04.csv
Download completed: raw_data/2025_04_zensu_en.csv
Downloading data for week 5 from URL: https://id-info.jihs.go.jp/surveillance/idwr/en/rapid/2025/05/zensu05.csv
Download completed: raw_data/2025_05_zensu_en.csv
All requested weeks have been processed.
Warning messages:
1: In download.file(url, destfile = dest_file, mode = "wb") :
URL https://id-info.jihs.go.jp/surveillance/idwr/en/rapid/2025/01/zensu01.csv: cannot open destfile 'raw_data/2025_01_zensu_en.csv', reason 'No such file or directory'
2: In download.file(url, destfile = dest_file, mode = "wb") :
download had nonzero exit status
3: In download.file(url, destfile = dest_file, mode = "wb") :
URL https://id-info.jihs.go.jp/surveillance/idwr/en/rapid/2025/02/zensu02.csv: cannot open destfile 'raw_data/2025_02_zensu_en.csv', reason 'No such file or directory'
4: In download.file(url, destfile = dest_file, mode = "wb") :
download had nonzero exit status
5: In download.file(url, destfile = dest_file, mode = "wb") :
URL https://id-info.jihs.go.jp/surveillance/idwr/en/rapid/2025/03/zensu03.csv: cannot open destfile 'raw_data/2025_03_zensu_en.csv', reason 'No such file or directory'
6: In download.file(url, destfile = dest_file, mode = "wb") :
download had nonzero exit status
7: In download.file(url, destfile = dest_file, mode = "wb") :
URL https://id-info.jihs.go.jp/surveillance/idwr/en/rapid/2025/04/zensu04.csv: cannot open destfile 'raw_data/2025_04_zensu_en.csv', reason 'No such file or directory'
8: In download.file(url, destfile = dest_file, mode = "wb") :
download had nonzero exit status
9: In download.file(url, destfile = dest_file, mode = "wb") :
URL https://id-info.jihs.go.jp/surveillance/idwr/en/rapid/2025/05/zensu05.csv: cannot open destfile 'raw_data/2025_05_zensu_en.csv', reason 'No such file or directory'
10: In download.file(url, destfile = dest_file, mode = "wb") :
download had nonzero exit status
...
All requested weeks have been processed.
There were 50 or more warnings (use warnings() to see the first 50)
```

it may be worth clarifying these or if this is the result of the target directory,
i.e. 'raw_data' not found to either prompt for creating it or stop the execution until the directory is found.
This should also be clarified in the documentation, i.e. README file.

Even when the target directory exist, I'm still seeing some warnings:
```
Warning messages:
1: In download.file(url, destfile = dest_file, mode = "wb") :
downloaded length 16070 != reported length 2522165
2: In download.file(url, destfile = dest_file, mode = "wb") :
URL 'https://id-info.jihs.go.jp/niid/images/idwr/ydata/2020/Syuukei/Syu_01_1.xlsx': status was 'Failed writing received data to disk/application'
3: In value[[3L]](cond) :
Download failed: https://id-info.jihs.go.jp/niid/images/idwr/ydata/2020/Syuukei/Syu_01_1.xlsx
Please check the server status or try again later.
4: In download.file(url, destfile = dest_file, mode = "wb") :
downloaded length 16070 != reported length 2470415
5: In download.file(url, destfile = dest_file, mode = "wb") :
URL 'https://id-info.jihs.go.jp/surveillance/idwr/annual/2021/syulist/Syu_01_1.xlsx': status was 'Failed writing received data to disk/application'
6: In value[[3L]](cond) :
...
```

# Data Import examples
I'm also running into problems with the examples listed in the "Data Import" session:

I believe the "raw_data" directory should be specified in the arguments of the function `jpinfect_read_confirmed` otherwise is failing, and even if I do that it still fails:
```
## Import bullet case reports
# Import all English reports in a directory
bullet <- jpinfect_read_bullet(directory = "raw_data")

# Import specific period of the data for 2025, weeks 1 to 10
bullet2025 <- jpinfect_read_bullet(year = 2025, week = 1:10, directory = "raw_data")
Error in jpinfect_read_confirmed(path = "2021_Syu_01_1.xlsx") :
Invalid path: must be either a file or directory
> dataset2021 <- jpinfect_read_confirmed(path = "raw_data/2021_Syu_01_1.xlsx")
Processing...Warning: Caught simpleError. Canceling all iterations ...
Error in utils::unzip(zip_path, list = TRUE) :
zip file '/.../raw_data/2021_Syu_01_1.xlsx' cannot be opened
```

Even if I change the directory manually before issuing the commands, still fails with a bizarre error as the file is present:
```
setwd("raw_data/")
> dataset2021 <- jpinfect_read_confirmed(path = "2021_Syu_01_1.xlsx")
Processing...
Warning: Caught simpleError. Canceling all iterations ...
Error: `path` does not exist: ‘2021_Syu_01_1.xlsx’
>
> dir()
[1] "2001_Syu_02_1.xls"    "2002_Syu_02_1.xls"    "2003_Syu_02_1.xls"
[4] "2004_Syu_02_1.xls"    "2005_Syu_02_1.xls"    "2006_Syu_02_1.xls"
[7] "2007_Syu_02_1.xls"    "2008_Syu_02_1.xls"    "2009_Syu_02_1.xls"
[10] "2010_Syu_02_1.xls"    "2011_Syu_02_1.xls"    "2012_Syu_02_1.xls"
[13] "2013_Syu_02_1.xls"    "2014_Syu_02_1.xlsx"   "2015_Syu_02_1.xlsx"
[16] "2016_Syu_02_1.xlsx"   "2017_Syu_02_1.xlsx"   "2018_Syu_02_1.xlsx"
[19] "2019_Syu_02_1.xlsx"   "2020_Syu_01_1.xlsx"   "2020_Syu_02_1.xlsx"
[22] "2021_Syu_01_1.xlsx"   "2021_Syu_02_1.xlsx"   "2022_Syu_02_1.xlsx"
[25] "2023_Syu_02_1.xlsx"   "2025_01_zensu_en.csv" "2025_01_zensu_jp.csv"
[28] "2025_02_zensu_en.csv" "2025_02_zensu_jp.csv" "2025_03_zensu_en.csv"
[31] "2025_03_zensu_jp.csv" "2025_04_zensu_en.csv" "2025_04_zensu_jp.csv"
[34] "2025_05_zensu_en.csv" "2025_05_zensu_jp.csv" "2025_06_zensu_jp.csv"
[37] "2025_07_zensu_jp.csv" "2025_08_zensu_jp.csv" "2025_09_zensu_jp.csv"
[40] "2025_10_zensu_jp.csv" "2025_11_zensu_jp.csv" "2025_12_zensu_jp.csv"
[43] "2025_13_zensu_jp.csv" "2025_14_zensu_jp.csv" "2025_15_zensu_jp.csv"
[46] "2025_16_zensu_jp.csv" "2025_17_zensu_jp.csv" "2025_18_zensu_jp.csv"
[49] "2025_19_zensu_jp.csv" "2025_20_zensu_jp.csv" "2025_21_zensu_jp.csv"
[52] "2025_22_zensu_jp.csv" "2025_23_zensu_jp.csv" "2025_24_zensu_jp.csv"
[55] "2025_25_zensu_jp.csv" "2025_26_zensu_jp.csv" "2025_27_zensu_jp.csv"
[58] "2025_28_zensu_jp.csv" "2025_29_zensu_jp.csv" "2025_30_zensu_jp.csv"
[61] "2025_31_zensu_jp.csv" "2025_32_zensu_jp.csv" "2025_33_zensu_jp.csv"
[64] "2025_34_zensu_jp.csv" "2025_35_zensu_jp.csv" "2025_36_zensu_jp.csv"
>
```
