### 2025-09-20 3c1017a
Stop if dest_dir was not created by users.
