### 2025-09-20 579ccef
File size checks were added to filter failed and incomplete downloads.
