### 2025-09-20 TomonoriHoshi/jpinfect#6 @TomonoriHoshi
@mponce0
Thank you very much for pointing out these issues.

I’ve added the following explanations to the README to guide users to manually create the destination folder before running the functions.

> The downloaded files are saved under the specified directory (e.g., raw_data folder).
> Please ensure that the destination directory exists prior to execution, as the function does not create directories automatically. This behaviour complies with CRAN policy, which prohibits packages from creating directories within the user's environment.
> If no destination is specified, the datasets will be stored in a temporary folder. Please note that this folder is session-specific and will be deleted once the R session is closed.

Additionally, I’ve updated both `jpinfect_read_bullet` and `jpinfect_read_confirmed` so that they now stop with an informative error message when the specified folder does not exist. [Commit details](https://github.com/TomonoriHoshi/jpinfect/commit/3c1017abf062552d4046d18733578b1281b78820)

Moreover, the error message stating "Please check the server status or try again later." is most likely due to the government server being temporarily unavailable or unresponsive.  According to the errors, it seems that some of the files could have not been completed their download. Usually, xls and xlsx files are > 2MB.  So, I added file size check functions to `jpinfect_read_confirmed`.  If the function detects the file size error, users are advised to re-download the file. [Commit details. ](https://github.com/TomonoriHoshi/jpinfect/commit/579ccef1e328fa5040931c0f522d25af76467c91)

Please kindly check these updates by installing the GitHub version of the package, and let me know if the errors are resolved.

```
if(!require("remotes")) install.packages("remotes")
remotes::install_github("TomonoriHoshi/jpinfect")
```
