### 2025-09-22 TomonoriHoshi/jpinfect#6 @mponce0
Still seeing errors when running,

```
# Download data for 2020 and 2021
jpinfect_get_confirmed(years = c(2020, 2021), type = "sex", dest_dir = "raw_data")

# Download data for all available years
jpinfect_get_confirmed(type = "place", dest_dir = "raw_data")

# Download English data for weeks 1 to 5 in 2025
jpinfect_get_bullet(year = 2025, week = 1:5, dest_dir = "raw_data")

# Download Japanese data for all weeks in 2025
jpinfect_get_bullet(year = 2025, language = "jp", dest_dir = "raw_data")
```

still seeing,

```
Warning messages:
1: In download.file(url, destfile = dest_file, mode = "wb") :
downloaded length 16006 != reported length 2522165
2: In download.file(url, destfile = dest_file, mode = "wb") :
URL 'https://id-info.jihs.go.jp/niid/images/idwr/ydata/2020/Syuukei/Syu_01_1.xlsx': status was 'Failedwriting received data to disk/application'
3: In value[[3L]](cond) :
Download failed: https://id-info.jihs.go.jp/niid/images/idwr/ydata/2020/Syuukei/Syu_01_1.xlsx
Please check the server status or try again later.
4: In download.file(url, destfile = dest_file, mode = "wb") :
downloaded length 16070 != reported length 2470415
5: In download.file(url, destfile = dest_file, mode = "wb") :
URL 'https://id-info.jihs.go.jp/surveillance/idwr/annual/2021/syulist/Syu_01_1.xlsx': status was 'Failed writing received data to disk/application'
6: In value[[3L]](cond) :
Download failed: https://id-info.jihs.go.jp/surveillance/idwr/annual/2021/syulist/Syu_01_1.xlsx
Please check the server status or try again later.
Downloading: https://idsc.niid.go.jp/idwr/CDROM/Kako/H13/Syuukei/Syu_02_1.xls
trying URL 'https://idsc.niid.go.jp/idwr/CDROM/Kako/H13/Syuukei/Syu_02_1.xls'
Content type 'application/vnd.ms-excel' length 3512832 bytes (3.4 MB)
```

because of this --I'd imagine--, the following example is also failing,

```
> dataset2021 <- jpinfect_read_confirmed(path = "2021_Syu_01_1.xlsx")
Error in jpinfect_read_confirmed(path = "2021_Syu_01_1.xlsx") :
Invalid path: must be either a file or directory
```

```
place_dataset <- jpinfect_read_confirmed(path = "raw_data", type = "place")

## Import bullet case reports
# Import all English reports in a directory
Error in .jpinfect_read_excels(type = type, directory = path, ...) :
The following files are unusually small (< 2000 KB) and may be corrupted:
raw_data/2001_Syu_02_1.xls
raw_data/2002_Syu_02_1.xls
raw_data/2003_Syu_02_1.xls
raw_data/2004_Syu_02_1.xls
raw_data/2005_Syu_02_1.xls
raw_data/2006_Syu_02_1.xls
raw_data/2007_Syu_02_1.xls
raw_data/2008_Syu_02_1.xls
raw_data/2009_Syu_02_1.xls
raw_data/2010_Syu_02_1.xls
raw_data/2011_Syu_02_1.xls
raw_data/2012_Syu_02_1.xls
raw_data/2013_Syu_02_1.xls
raw_data/2014_Syu_02_1.xlsx
raw_data/2015_Syu_02_1.xlsx
raw_data/2016_Syu_02_1.xlsx
raw_data/2017_Syu_02_1.xlsx
raw_data/2018_Syu_02_1.xlsx
raw_data/2019_Syu_02_1.xlsx
raw_data/2020_Syu_02_1.xlsx
raw_data/2021_Syu_02_1.xlsx
raw_data/2022_Syu_02_1.xlsx
raw_data/2023_Syu_02_1.xlsx
This may indicate a failed or incomplete download.

Please re-run jpinfect_get_confirmed() to re-download them.
```
