### 2025-09-24 7514406
Improved error message for invald path.
