### 2025-09-24 2f9e1c4
Improved raw_data folder description and fixed an error in the example to import a signle without raw_data path.
