### 2025-09-24 4402d52
Enhance Excel download robustness: retry up to 3 times on failure or HTML error page detection.
