### 2025-09-24 d05fe89
Enhance Excel download robustness: retry up to 3 times on failure or HTML error page detection with multiple hmtl_signatures.
