### 2025-09-25 7486305
Enhance Excel download robustness: file size check and timeout extended.
