### 2025-09-25 6e4c0db
Enhance Excel download robustness.
