+++ joss/paper.bib
@@ -5,11 +5,12 @@
     address = {Vienna, Austria},
     year = {2025},
     url = {https://www.R-project.org/},
+    doi = {10.32614/r.manuals}
 }
 
 @book{Vynnycky:2011,
 	location = {Oxford},
-	edition = {reprint},
+	edition = {1st},
 	title = {An introduction to infectious disease modelling},
 	isbn = {978-0-19-856576-5},
 	pagetotal = {370},
