### 2025-09-25 #8577 @fboehm
> [@TomonoriHoshi](https://github.com/TomonoriHoshi) [@fboehm](https://github.com/fboehm) -- update: currently my only points are related to the issue raised in [TomonoriHoshi/jpinfect#6 (comment)](https://github.com/TomonoriHoshi/jpinfect/issues/6#issuecomment-3321344458)
>
> which relates also to the unchecked items in my checklist.
>
> Please also notice the references issues just above.

@TomonoriHoshi - please fix the reference issues noted above. Thank you, @mponce0 !
