### 2025-09-25 TomonoriHoshi/jpinfect#6 @FIELDWORKERS
Thanks again.  I identified two separate issues:

1) Path correction
The correct path should include "raw_data" as shown below:
`dataset2021 <- jpinfect_read_confirmed(path = "raw_data/2021_Syu_01_1.xlsx")`
I've updated the README accordingly and added some notes about the `raw_data` folder used in the examples: [Commits](https://github.com/TomonoriHoshi/jpinfect/commit/d05fe89e336a229c2bf619578899e540d25b9adc).

2) Errors when reading Excel files.
It seems the `jpinfect_get_confirmed` was unable to successfully retrieve data from the JIHS server. I've revised the code to improve the robustness of the function.  Please try the updated one:
`jpinfect_get_bullet(year = 2025, language = "jp", dest_dir = "raw_data")`
to retrieve the data from the server.  [Commits] (https://github.com/TomonoriHoshi/jpinfect/commit/6e4c0db1913227a8ebcd6f338816c1e962b7b6da)
