### 2025-09-26 TomonoriHoshi/jpinfect#6 @mponce0
I'm still seeing errors --even when re-installing the package from the master branch of the repository--:

```
> # Download data for 2020 and 2021
jpinfect_get_confirmed(years = c(2020, 2021), type = "sex", dest_dir = "raw_data")
Downloading: https://id-info.jihs.go.jp/niid/images/idwr/ydata/2020/Syuukei/Syu_01_1.xlsx
Error in download.file(url, destfile = dest_file, mode = "wb", quiet = TRUE) :
download from 'https://id-info.jihs.go.jp/niid/images/idwr/ydata/2020/Syuukei/Syu_01_1.xlsx' failed
```

If the issue is related to _unavailability_ of the **remote** data sets, I'd recommend looking into the possibility of storing "local" snapshots of these (or some critical) datasets within your package.
We followed an strategy like this in https://doi.org/10.21105/joss.02995.
This strategy also mitigates connectivity issues and/or allows users to access snapshoted datasets when they are not connected online. The caveat is that you will need to warn the user about the "life-time" of the dataset, i.e. when it was updated/accessed, etc.
