### 2025-09-26 TomonoriHoshi/jpinfect#6 @TomonoriHoshi
Thank you for the check and suggestion.  May I ask if you are able to download the data from the following link manually? -  [https://id-info.jihs.go.jp/niid/images/idwr/ydata/2020/Syuukei/Syu_01_1.xlsx](https://id-info.jihs.go.jp/niid/images/idwr/ydata/2020/Syuukei/Syu_01_1.xlsx)
