### 2025-09-27 TomonoriHoshi/jpinfect#6 @mponce0
I was able to run the "Data Acquisition" example to completion, as well as, to fetch the data directly from the URL.
But this indicates to me that there may be issues with the reliability from where the data is being hosted.
Furthermore, my assumption would be that you don't control this resource hence if the sources are unreachable (at it has been several times for me) then you need a mechanism to either provide a gracious exit or as I suggested before trigger the utilization of local available data provided by the package.

The "Data Import" example is taking a really long time, I noticed the message in the standard output but this should be also mentioned in the documentation to warn the user about it before running them.
