### 2025-09-27 TomonoriHoshi/jpinfect#7 @mponce0
test cases failing
When running `devtools::test()` I see 5 test cases failing:

```
── Failed tests ───────────────────────────────────────────────────────────────────────────────────────────────────
Error (test-jpinfect_read_confirmed.R:10:3): .jpinfect_read_excel imports Excel file correctly
Error in `.jpinfect_read_excel(file_path = temp_file, year = 2014, sheet_range = 3)`: The file size is unusually small (< 2000 KB) and may be corrupted.
This may indicate a failed or incomplete download.
Please re-run jpinfect_get_confirmed() to re-download the file.
Backtrace:
▆
1. └─jpinfect:::.jpinfect_read_excel(...) at test-jpinfect_read_confirmed.R:10:3

Error (test-jpinfect_read_confirmed.R:20:3): .jpinfect_read_excel handles errors appropriately
Error in `.jpinfect_read_excel(file_path = unsupported_file)`: The file size is unusually small (< 2000 KB) and maybe corrupted.
This may indicate a failed or incomplete download.
Please re-run jpinfect_get_confirmed() to re-download the file.
Backtrace:
▆
1. ├─testthat::expect_error(...) at test-jpinfect_read_confirmed.R:20:3
2. │ └─testthat:::expect_condition_matching(...)
3. │   └─testthat:::quasi_capture(...)
4. │     ├─testthat (local) .capture(...)
5. │     │ └─base::withCallingHandlers(...)
6. │     └─rlang::eval_bare(quo_get_expr(.quo), quo_get_env(.quo))
7. └─jpinfect:::.jpinfect_read_excel(file_path = unsupported_file)

Error (test-jpinfect_read_confirmed.R:54:3): .jpinfect_read_excels processes sex data correctly (skip_on_cran)
Error in `.jpinfect_read_excels(type = "sex", directory = temp_data_dir)`: The following files are unusually small (< 2000 KB) and may be corrupted:
/home/l/lcl_uotcscd71/lcl_uotcscd71s1402/jpinfect/inst/extdata/1999_Syu_11.xls
/home/l/lcl_uotcscd71/lcl_uotcscd71s1402/jpinfect/inst/extdata/2014_Syu_01_1.xlsx
This may indicate a failed or incomplete download.

Please re-run jpinfect_get_confirmed() to re-download them.

Backtrace:
▆
1. └─jpinfect:::.jpinfect_read_excels(type = "sex", directory = temp_data_dir) at test-jpinfect_read_confirmed.R:54:3

Error (test-jpinfect_read_confirmed.R:64:3): .jpinfect_read_excels processes place data correctly (skip_on_cran)
Error in `.jpinfect_read_excels(type = "place", directory = temp_data_dir)`: The following files are unusually small (< 2000 KB) and may be corrupted:
/home/l/lcl_uotcscd71/lcl_uotcscd71s1402/jpinfect/inst/extdata/2014_Syu_02_1.xlsx
/home/l/lcl_uotcscd71/lcl_uotcscd71s1402/jpinfect/inst/extdata/2017_Syu_02_1.xlsx
This may indicate a failed or incomplete download.

Please re-run jpinfect_get_confirmed() to re-download them.

Backtrace:
▆
1. └─jpinfect:::.jpinfect_read_excels(type = "place", directory = temp_data_dir) at test-jpinfect_read_confirmed.R:64:3

Error (test-jpinfect_read_confirmed.R:84:3): jpinfect_read_confirmed validates arguments
Error in `jpinfect_read_confirmed(path = "path/to/file.xls")`: Invalid path: no file or directory could be located.Please verify that the specified path is correct.
Backtrace:
▆
1. ├─testthat::expect_error(...) at test-jpinfect_read_confirmed.R:84:3
2. │ └─testthat:::expect_condition_matching(...)
3. │   └─testthat:::quasi_capture(...)
4. │     ├─testthat (local) .capture(...)
5. │     │ └─base::withCallingHandlers(...)
6. │     └─rlang::eval_bare(quo_get_expr(.quo), quo_get_env(.quo))
7. └─jpinfect::jpinfect_read_confirmed(path = "path/to/file.xls")

[ FAIL 5 | WARN 0 | SKIP 0 | PASS 71 ]
>
```
