### 2025-09-30 TomonoriHoshi/jpinfect#7 @TomonoriHoshi
Thank you very much for reporting this.  Please kindly check the updated version.  Here are [commit details](https://github.com/TomonoriHoshi/jpinfect/commit/b45af4b36f7eec568b1ee1a9d090a99cdc4e5b21).
