### 2025-09-30 add0560
Clarify download failure handling and suggest Posit Cloud as alternative; added notes about required time to import dataset
