### 2025-09-30 fe7a5f1
Improved download error handling and message output
