### 2025-10-01 TomonoriHoshi/jpinfect#6 @TomonoriHoshi
Thank you very much for testing in the cloud server and for your suggestions.

The [latest commits](https://github.com/TomonoriHoshi/jpinfect/commit/fe7a5f11f754274e19973442e95d57dba3db7f60) have improved error message output as follows:
```
> jpinfect_get_confirmed(years = 2002, type = "place", dest_dir = "raw_data")
Downloading: https://idsc.niid.go.jp/idwr/CDROM/Kako/H14/Syuukei/Syu_02_1.xls
Download failed due to connection error.
Retrying download (1/3)... (status=1)
Download failed due to connection error.
Retrying download (2/3)... (status=1)
Download failed due to connection error.
Retrying download (3/3)... (status=1)
Download failed: https://idsc.niid.go.jp/idwr/CDROM/Kako/H14/Syuukei/Syu_02_1.xls
Please check the server status or try again later.
If the issue persists, consider using built-in datasets:
  - sex_prefecture
  - place_prefecture
  - bullet
```

Also, the updated README mentioned required time for data import example ([commit details](https://github.com/TomonoriHoshi/jpinfect/commit/add056069231ec82be98265f48a2d827e40b368d)):

> ⚠️Note: The import process may require several minutes to complete, particularly when processing large directories. This example may take ~10 minutes.

I believe the latest changes address the issue. If you're happy with the outcome, please feel free to close this.
