### 2025-10-02 8690a51
Add R-CMD-check workflow for main/master branches
