### 2025-10-02 21932fd
Add R-CMD-check workflow for main/master branches
