### 2025-10-02 f4b2cbb
Update README.md
Remove merge conflict mannualy
