### 2025-10-02 c442207
Fixed CI run error for skipping size check
