### 2025-10-02 c2e9286
Fix CI error for file size check - Line 104
