+++ R/jpinfect_read_confirmed.R
@@ -101,7 +101,7 @@ jpinfect_read_confirmed <- function(path, type = NULL, ...) {
   }
 
   # file size check
-  if (!nzchar(Sys.getenv("JPINFECT_SKIP_SIZE_CHECK") || Sys.getenv("CI") == "true")) {
+  if (!(nzchar(Sys.getenv("JPINFECT_SKIP_SIZE_CHECK")) || Sys.getenv("CI") == "true")) {
     file_size <- file.info(file_path)$size
     if (!is.na(file_size) && file_size < 2000000) {
       stop(paste("The file size is unusually small (< 2000 KB) and may be corrupted.\n",
