### 2025-10-04 TomonoriHoshi/jpinfect#6 @mponce0
Examples are working, only thing I'm noticing now is that the plot generated in the last example, after following all the previous steps is different than the one shown and this could be confusing to the user.
I.e. the plot I was able to generate produces a histogram only for 2025-01 while the one displayed in the README is for whole range of years.
For example, I am able to generate the precise plot as shown but without following the whole sequence of examples, so is possible that the plot presented in the documentation was generated in an independent example but again this should be clearly stated for clarity sake otherwise the user may think something is off in terms of "reprehensibility" of expected results or if variations are expected then again this should be clearly disclosed.
