### 2025-10-05 bc3798d
Changes to follow suggestions by @mponce0: 1) "10 mins" to "a substantiaon amount of time..." 2) dataset name from bullet to bullet_local for the example 3) Added notes for plot difference
