### 2025-10-05 TomonoriHoshi/jpinfect#6 @TomonoriHoshi
Thank you very much for taking the time to review the codes.

To avoid confusion, I've updated the dataset name used in the example - from  `bullet` to `bullet_local`: [commit details](https://github.com/TomonoriHoshi/jpinfect/commit/bc3798d174c6e6e7eb20f53ffc8c57c5034c588a#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L154). This ensures that users can now reproduce the sample plot shown in the example.

Additionally, I’ve added a note to clarify that the example plot may become outdated due to delays in updating plots following changes to the built-in dataset. The updated REAME now includes the following:

> Note: The example plot reflects a snapshot of the dataset at the time of writing. Due to periodic updates to built-in data, your results may vary slightly.

I hope these updates covered your points. If you're happy with it, you're welcome to close the thread.
