### 2025-10-05 #8577 @JDRomano2
Fully agreed with @mponce0 , and I have no further recommendations to add. I'd like to confirm that things work smoothly with the new CRAN release, and once that is completed my review will be "complete".
