### 2025-10-06 6d680c1
A minior label adjustment for the example plot.
