### 2025-10-07 #8577 @TomonoriHoshi
@JDRomano2 Thank you very much for your comments.  I've just uploaded [version 0.1.4](https://cran.r-project.org/web/packages/jpinfect/index.html) to CRAN. Currently, only the source package is available; I expect the binaries to be ready in a couple of days.

@mponce0 Many thanks for closing the issues and completing all checklist items.  Much appreciated.
