### 2025-10-07 52898cd
Fixed issues detected on CRAN: 1) stat verson removed 2) file size check skipped on CRAN, for test sample file.
