+++ DESCRIPTION
@@ -37,7 +37,7 @@ Imports:
     magrittr (>= 2.0.3),
     readr (>= 2.1.5),
     readxl (>= 1.4.5),
-    stats (>= 4.4.3),
+    stats,
     stringi (>= 1.8.7),
     stringr (>= 1.5.1),
     tidyr (>= 1.3.1),
