+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: jpinfect
 Type: Package
 Title: Acquiring and Processing Data from Japan Institute for Health Security
-Version: 0.1.4
+Version: 0.1.5
 Authors@R: c(
     person(
       "Tomonori", "Hoshi",
