### 2025-10-08 #8577 @TomonoriHoshi
@JDRomano2 The CRAN test for version 0.1.4 failed because I forgot to disable the 2MB file size check function during CRAN testing—[this was the same issue previously reported on a local machine](https://github.com/TomonoriHoshi/jpinfect/issues/7). I've already submitted version 0.1.5 with the fix, but the 0.1.4 binaries are still being built across all OSs. Once 0.1.4 completes, I expect 0.1.5 to follow shortly.

Hopefully, version 0.1.5 will be available on CRAN soon for testing.
