### 2025-11-09 c8917ab
Fix title formatting in Hanscombe 2019 citation
