+++ joss/paper.bib
@@ -125,7 +125,7 @@
 
 
 @article{Hanscombe:2019,
-    title = {ukbtools: An R package to manage and query UK Biobank data},
+    title = {{ukbtools: An R package to manage and query UK Biobank data}},
     author = {Hanscombe, Ken B. AND Coleman, Jonathan R. I. AND Traylor, Matthew AND Lewis, Cathryn M.},
     year = {2019},
     month = {05},
