### 2025-11-10 #8577 @TomonoriHoshi
@fboehm Thanks for checking.  I've just updated the reference issue.  My apologies, but I realised that the correct version is  0.1.6 (not v0.1.6).
