### 2025-11-13 #8577 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
