### 2025-11-16 #8577 @Kevin-Mattheus-Moerman
👉 Please create a release with the tag `0.1.6` for you repository so it is listed here: https://github.com/TomonoriHoshi/jpinfect/tags
